fix(utils): move some function to utils
