+++ codecept.conf.js
@@ -42,6 +42,7 @@ exports.config = {
     users: './utils/user.js',
     google: './utils/google.js',
     files: './utils/file.js',
+    dataUploadUtil: './utils/dataUploadUtil',
 
     // APIs
     sheepdog: './services/apis/sheepdog/sheepdogService.js',
