+++ suites/apis/dataUploadTest.js
@@ -43,7 +43,7 @@ const getUploadUrlFromFence = async function (fence, users, indexd) {
  * - Link metadata to the file via sheepdog
  * - Download the file via fence and check who can download and when
  */
-Scenario('File upload and download via API calls', async (fence, users, nodes, indexd, sheepdog, files) => {
+Scenario('File upload and download via API calls', async (fence, users, nodes, indexd, sheepdog, dataUploadUtil) => {
   // request a  presigned URL from fence
   let fenceUploadRes = await getUploadUrlFromFence(fence, users, indexd);
   let fileGuid = fenceUploadRes.body.guid;
@@ -65,7 +65,7 @@ Scenario('File upload and download via API calls', async (fence, users, nodes, i
   // we simulate not waiting by not uploading the file to the S3 bucket yet
 
   // fail to submit metadata for this file
-  let sheepdogRes = await files.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
+  let sheepdogRes = await nodes.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
   sheepdog.ask.hasStatusCode(sheepdogRes.addRes, 400, 'Linking metadata to file without hash and size should not be possible');
 
   // check that we CANNOT download the file (there is no URL in indexd yet)
@@ -73,10 +73,10 @@ Scenario('File upload and download via API calls', async (fence, users, nodes, i
   fence.ask.hasNoUrl(signedUrlRes);
 
   // now, upload the file to the S3 bucket using the presigned URL
-  await files.uploadFileToS3(presignedUrl, filePath, fileSize);
+  await dataUploadUtil.uploadFileToS3(presignedUrl, filePath, fileSize);
 
   // wait for the indexd listener to add size, hashes and URL to the record
-  await files.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
+  await dataUploadUtil.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
 
   // Try downloading before linking metadata to the file. It should succeed
   // for the uploader but fail for other users
@@ -93,7 +93,7 @@ Scenario('File upload and download via API calls', async (fence, users, nodes, i
   fence.ask.assertStatusCode(signedUrlRes, 401, 'User who is not the uploader should not successfully download file before metadata linking');
 
   // submit metadata for this file
-  sheepdogRes = await files.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
+  sheepdogRes = await nodes.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
   sheepdog.ask.addNodeSuccess(sheepdogRes);
 
   // a user who is not the uploader can now download the file
@@ -105,7 +105,7 @@ Scenario('File upload and download via API calls', async (fence, users, nodes, i
 
   // check that we cannot link metadata to a file that already has metadata:
   // try to submit metadata for this file again
-  sheepdogRes = await files.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5, submitter_id='submitter_id_new_value');
+  sheepdogRes = await nodes.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5, submitter_id='submitter_id_new_value');
   sheepdog.ask.addNodeSuccess(sheepdogRes);
 });
 
@@ -129,7 +129,7 @@ Scenario('User without role cannot upload', async (fence, users, nodes, indexd) 
 /**
  * This time, use the gen3 data client to upload and download the file
  */
-Scenario('File upload and download via client', async (dataClient, indexd, nodes, files) => {
+Scenario('File upload and download via client', async (dataClient, indexd, nodes, files, dataUploadUtil) => {
   // use gen3 client to upload a file
   let fileGuid = await dataClient.do.uploadFile(filePath);
   createdGuids.push(fileGuid);
@@ -145,7 +145,7 @@ Scenario('File upload and download via client', async (dataClient, indexd, nodes
   await indexd.complete.checkRecordExists(fileNode);
 
   // wait for the indexd listener to add size, hashes and URL to the record
-  await files.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
+  await dataUploadUtil.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
 
   // download the file via the data client
   downloadPath = './tmpFileDestination.txt';
@@ -156,14 +156,14 @@ Scenario('File upload and download via client', async (dataClient, indexd, nodes
  * Upload a file, then delete it through fence. Aftert deletion, the file
  * should not be accessible for metadata linking or download
  */
-Scenario('Data file deletion', async (fence, users, indexd, sheepdog, nodes, files) => {
+Scenario('Data file deletion', async (fence, users, indexd, sheepdog, nodes, dataUploadUtil) => {
   // request a  presigned URL from fence
   let fenceUploadRes = await getUploadUrlFromFence(fence, users, indexd);
   let fileGuid = fenceUploadRes.body.guid;
   let presignedUrl = fenceUploadRes.body.url;
 
   // upload the file to the S3 bucket using the presigned URL
-  await files.uploadFileToS3(presignedUrl, filePath, fileSize);
+  await dataUploadUtil.uploadFileToS3(presignedUrl, filePath, fileSize);
 
   fileNode = {
     did: fileGuid,
@@ -174,7 +174,7 @@ Scenario('Data file deletion', async (fence, users, indexd, sheepdog, nodes, fil
   };
 
   // wait for the indexd listener to add size, hashes and URL to the record
-  await files.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
+  await dataUploadUtil.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
 
   // check that a user who is not the uploader cannot delete the file
   let fenceRes = await fence.do.deleteFile(fileGuid, userHeader=users.auxAcct1.accessTokenHeader);
@@ -188,7 +188,7 @@ Scenario('Data file deletion', async (fence, users, indexd, sheepdog, nodes, fil
   indexd.ask.resultFailure(indexdRes);
 
   // no metadata linking after delete
-  let sheepdogRes = await files.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
+  let sheepdogRes = await nodes.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
   sheepdog.ask.hasStatusCode(sheepdogRes.addRes, 400, 'Metadata linking should not be possible after file deletion');
 
   // no download after delete
@@ -200,7 +200,7 @@ Scenario('Data file deletion', async (fence, users, indexd, sheepdog, nodes, fil
  * Upload 2 files with the same contents (so same hash and size) and
  * link metadata to them via sheepdog
  */
-Scenario('Upload the same file twice', async (sheepdog, indexd, nodes, users, fence, files) => {
+Scenario('Upload the same file twice', async (sheepdog, indexd, nodes, users, fence, dataUploadUtil) => {
   ////////////
   // FILE 1 //
   ////////////
@@ -212,7 +212,7 @@ Scenario('Upload the same file twice', async (sheepdog, indexd, nodes, users, fe
   let presignedUrl = fenceUploadRes.body.url;
 
   // upload the file to the S3 bucket using the presigned URL
-  await files.uploadFileToS3(presignedUrl, filePath, fileSize);
+  await dataUploadUtil.uploadFileToS3(presignedUrl, filePath, fileSize);
 
   let fileNode = {
     did: fileGuid,
@@ -223,10 +223,10 @@ Scenario('Upload the same file twice', async (sheepdog, indexd, nodes, users, fe
   };
 
   // wait for the indexd listener to add size, hashes and URL to the record
-  await files.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
+  await dataUploadUtil.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
 
   // submit metadata for this file
-  let sheepdogRes = await files.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
+  let sheepdogRes = await nodes.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5);
   sheepdog.ask.addNodeSuccess(sheepdogRes, 'first upload');
 
   // check that the file can be downloaded
@@ -247,14 +247,14 @@ Scenario('Upload the same file twice', async (sheepdog, indexd, nodes, users, fe
   presignedUrl = fenceUploadRes.body.url;
 
   // upload the file to the S3 bucket using the presigned URL
-  await files.uploadFileToS3(presignedUrl, filePath, fileSize);
+  await dataUploadUtil.uploadFileToS3(presignedUrl, filePath, fileSize);
 
   // wait for the indexd listener to add size, hashes and URL to the record
   fileNode.did = fileGuid;
-  await files.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
+  await dataUploadUtil.waitUploadFileUpdatedFromIndexdListener(indexd, fileNode);
 
   // submit metadata for this file
-  sheepdogRes = await files.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5, submitter_id='submitter_id_new_value');
+  sheepdogRes = await nodes.submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5, submitter_id='submitter_id_new_value');
   sheepdog.ask.addNodeSuccess(sheepdogRes, 'second upload');
 
   // check that the file can be downloaded
@@ -283,7 +283,7 @@ BeforeSuite(async (dataClient, fence, users, sheepdog, indexd, files) => {
   fileMd5 = await files.getFileHash(filePath);
 });
 
-AfterSuite(async (files, indexd) => {
+AfterSuite(async (files, indexd, dataUploadUtil) => {
   // delete the temp file from local storage
   if (fs.existsSync(filePath)) {
     files.deleteFile(filePath);
@@ -293,7 +293,7 @@ AfterSuite(async (files, indexd) => {
   // Note: we don't use fence's /delete endpoint here because it does not allow
   // deleting from indexd records that have already been linked to metadata
   await indexd.complete.deleteFiles(createdGuids);
-  await files.cleanS3(fileName, createdGuids);
+  await dataUploadUtil.cleanS3(fileName, createdGuids);
 });
 
 Before((nodes) => {
