+++ utils/file.js
@@ -91,113 +91,4 @@ module.exports = {
       fd.on('end', ()=>resolve(hash.read()));
     });
   },
-
-  /**
-   * upload a file to an S3 bucket using a presigned URL
-   */
-  async uploadFileToS3(presignedUrl, filePath, fileSize) {
-    fs.createReadStream(filePath).pipe(require('request')({
-      method: 'PUT',
-      url: presignedUrl,
-      headers: {
-        'Content-Length': fileSize
-      }
-    }, function (err, res, body) {
-      if (err) {
-        throw new Error(err);
-      }
-    }));
-  },
-
-  /** 
-   * create a file that contains the list of GUIDs of files that were uploaded  
-   * to s3 during this testing session. The files will be deleted during the  
-   * CleanS3 step of the Jenkins pipeline 
-   */ 
-  async cleanS3(fileName, createdGuids) {
-    if (inJenkins) {  
-      const dirName = `${homedir}/s3-cleanup`;  
-      if (!fs.existsSync(dirName)){ 
-        fs.mkdirSync(dirName);  
-      } 
-      this.createTmpFile(  
-        `${dirName}/${fileName}`, 
-        createdGuids.join("\n") 
-      );  
-      console.log(`Created ${fileName} in ${dirName}`); 
-    } 
-  },
-
-  /**
-   * wait until a file's hash and size are updated in indexd
-   */
-  async waitUploadFileUpdatedFromIndexdListener(indexd, fileNode) {
-    /**
-     * return true if the record has been updated in indexd, false otherwise
-     */
-    const isRecordUpdated = async function(indexd, fileNode) {
-      try {
-        // check if indexd was updated with the correct hash and size
-        await indexd.complete.checkFile(fileNode);
-        return true;
-      }
-      catch {
-        return false;
-      }
-    };
-
-    const timeout = 60; // max number of seconds to wait
-    let errorMessage = `The indexd listener did not complete the record after ${timeout} seconds`;
-
-    await smartWait(isRecordUpdated, [indexd, fileNode], timeout, errorMessage);
-  },
-
-  async generateAndAddCoremetadataNode(sheepdog) {
-    const newSubmitterID = `submitter-${Math.random().toString(36).substring(2, 7)}`;
-    const newCoremetadataNode = {
-      data: {
-        projects: {
-            code: 'jenkins',
-        },
-        submitter_id: newSubmitterID,
-        type: 'core_metadata_collection',
-      },
-    };
-    await sheepdog.complete.addNode(newCoremetadataNode);
-    return newSubmitterID;
-  },
-
-  /**
-   * link metadata to an indexd file via sheepdog,
-   * after submitting metadata for the parent nodes
-   * /!\ this function does not include a check for success or
-   * failure of the data_file node's submission
-   */
-  async submitGraphAndFileMetadata(sheepdog, nodes, fileGuid, fileSize, fileMd5, submitter_id=null) {
-    // submit metadata with object id via sheepdog
-    metadata = nodes.getFileNode().clone();
-    metadata.data.object_id = fileGuid;
-    metadata.data.file_size = fileSize;
-    metadata.data.md5sum = fileMd5;
-    if (submitter_id) {
-      metadata.data.submitter_id = submitter_id;
-    }
-
-    // assuming all data files can be submitted with a single link to a
-    // core_metadata_collection: add it, and remove other links
-    const cmcSubmitterID = await this.generateAndAddCoremetadataNode(sheepdog);
-    for (var prop in metadata.data) {
-      if (metadata.data.hasOwnProperty(prop) && metadata.data[prop].hasOwnProperty('submitter_id')) {
-        delete metadata.data[prop];
-      }
-    }
-    metadata.data.core_metadata_collections = {
-      submitter_id: cmcSubmitterID
-    };
-
-    await sheepdog.do.addNode(metadata); // submit, but don't check for success
-
-    // the result of the submission is stored in metadata.addRes by addNode()
-    return metadata;
-  },
 };
